+++ src/main/java/spms/controls/MemberDeleteController.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Component;
 
 import spms.bind.DataBinding;
 import spms.dao.MemberDao;
+import spms.vo.Member;
 
 //스프링 애노테이션(@Component)으로 변경
 @Component("/member/delete.do")
@@ -30,8 +31,12 @@ public class MemberDeleteController implements Controller, DataBinding {
   @Override
   public String execute(Map<String, Object> model) throws Exception {
     Integer no = (Integer)model.get("no");
-    memberDao.delete(no);
     HttpSession session = (HttpSession)model.get("session");
+    Member temp = (Member)session.getAttribute("member");
+    if(temp.getNo() != no)
+  	  return "redirect:list.do";
+    
+    memberDao.delete(no);
     session.setAttribute("member", null);
     return "redirect:list.do";
